+++ Source/VMware.vSphereDSC/VMware.vSphereDSC.psm1
@@ -8340,9 +8340,17 @@ class VMHostIScsiHba : VMHostIScsiHbaBaseDSC {
     [DscProperty(Key)]
     [string] $Name
 
-    hidden [string] $ConfigureIScsiHbaChapMessage = "Configuring CHAP settings of iSCSI Host Bus Adapter {0} from VMHost {1}."
+    <#
+    .DESCRIPTION
 
-    hidden [string] $CouldNotConfigureIScsiHbaChapMessage = "Could not configure CHAP settings of iSCSI Host Bus Adapter {0} from VMHost {1}. For more information: {2}"
+    Specifies the name for the VMHost Host Bus Adapter device.
+    #>
+    [DscProperty()]
+    [string] $IScsiName
+
+    hidden [string] $ConfigureIScsiHbaMessage = "Configuring iSCSI Host Bus Adapter {0} from VMHost {1}."
+
+    hidden [string] $CouldNotConfigureIScsiHbaMessage = "Could not configure iSCSI Host Bus Adapter {0} from VMHost {1}. For more information: {2}"
 
     [void] Set() {
         try {
@@ -8352,7 +8360,7 @@ class VMHostIScsiHba : VMHostIScsiHbaBaseDSC {
 
             $iScsiHba = $this.GetIScsiHba($this.Name)
 
-            $this.ConfigureIScsiHbaChap($iScsiHba)
+            $this.ConfigureIScsiHba($iScsiHba)
         }
         finally {
             $this.DisconnectVIServer()
@@ -8368,7 +8376,11 @@ class VMHostIScsiHba : VMHostIScsiHbaBaseDSC {
 
             $iScsiHba = $this.GetIScsiHba($this.Name)
 
-            $result = !$this.ShouldModifyCHAPSettings($iScsiHba.AuthenticationProperties)
+            $shouldConfigureiScsiHba = @(
+                $this.ShouldModifyCHAPSettings($iScsiHba.AuthenticationProperties),
+                $this.ShouldUpdateDscResourceSetting('IScsiName', $iScsiHba.IScsiName, $this.IScsiName)
+            )
+            $result = !($shouldConfigureiScsiHba -Contains $true)
 
             $this.WriteDscResourceState($result)
 
@@ -8403,9 +8415,9 @@ class VMHostIScsiHba : VMHostIScsiHbaBaseDSC {
     <#
     .DESCRIPTION
 
-    Configures the CHAP properties of the specified iSCSI Host Bus Adapter.
+    Configures the CHAP properties and the iScsi name of the specified iSCSI Host Bus Adapter.
     #>
-    [void] ConfigureIScsiHbaChap($iScsiHba) {
+    [void] ConfigureIScsiHba($iScsiHba) {
         $setVMHostHbaParams = @{
             IScsiHba = $iScsiHba
             Confirm = $false
@@ -8414,13 +8426,14 @@ class VMHostIScsiHba : VMHostIScsiHbaBaseDSC {
         }
 
         $this.PopulateCmdletParametersWithCHAPSettings($setVMHostHbaParams)
+        if (![string]::IsNullOrEmpty($this.IScsiName)) { $setVMHostHbaParams.IScsiName = $this.IScsiName }
 
         try {
-            Write-VerboseLog -Message $this.ConfigureIScsiHbaChapMessage -Arguments @($iScsiHba.Device, $this.VMHost.Name)
+            Write-VerboseLog -Message $this.ConfigureIScsiHbaMessage -Arguments @($iScsiHba.Device, $this.VMHost.Name)
             Set-VMHostHba @setVMHostHbaParams
         }
         catch {
-            throw ($this.CouldNotConfigureIScsiHbaChapMessage -f $iScsiHba.Device, $this.VMHost.Name, $_.Exception.Message)
+            throw ($this.CouldNotConfigureIScsiHbaMessage -f $iScsiHba.Device, $this.VMHost.Name, $_.Exception.Message)
         }
     }
 
@@ -8433,6 +8446,7 @@ class VMHostIScsiHba : VMHostIScsiHbaBaseDSC {
         $result.Server = $this.Connection.Name
         $result.VMHostName = $this.VMHost.Name
         $result.Name = $iScsiHba.Device
+        $result.IScsiName = $iScsiHba.IScsiName
         $result.ChapType = $iScsiHba.AuthenticationProperties.ChapType.ToString()
         $result.ChapName = [string] $iScsiHba.AuthenticationProperties.ChapName
         $result.MutualChapEnabled = $iScsiHba.AuthenticationProperties.MutualChapEnabled